feature: Allow video muting [sc-238963]
Add a new `shouldMute` parameter to YPVideoFiltersVC, which will allow
subclasses to toggle whether the output video will be muted or not.

The muting itself is super simple, we just skip adding the audio track to
the output. The only other change is basically to force the save flow to
go through `fetchVideoUrlAndCrop` even if there was no cropping/trimming
done, but the video should be muted.